% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memohelpR-package.R
\docType{package}
\name{memohelpR-package}
\alias{memohelpR}
\alias{memohelpR-package}
\title{memohelpR: Helps you memoise R functions with the memoise package}
\description{
memohelpR is an experimental R package that provides an easy way to memoize functions with a disk cache. The main function, memoise_fun_in_proj, creates a memoized version of a specified function, writing the cache to disk in a folder in the project.
}
\author{
\strong{Maintainer}: David Barnett \email{david.barnett@maastrichtuniversity.nl} (\href{https://orcid.org/0000-0003-1961-7206}{ORCID})

}
\keyword{internal}
