% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise_fun_in_proj.R
\name{memoise_fun_in_proj}
\alias{memoise_fun_in_proj}
\title{Memoise a Function in the Project Directory}
\usage{
memoise_fun_in_proj(
  func,
  omit_args = NULL,
  proj_dir = here::here(),
  memo_dir = "memoise",
  cache_dir = "cache",
  logfile = "log.txt",
  evict = "lru",
  ...
)
}
\arguments{
\item{func}{a function that needs to be memoised}

\item{omit_args}{Names of arguments to ignore when calculating hash}

\item{proj_dir}{the project directory where the memoised function and its cache will be stored}

\item{memo_dir}{the subdirectory within the project directory to store the memoised function}

\item{cache_dir}{the subdirectory within the memo_dir to store the cache}

\item{logfile}{the file name of the log file to store cache activities (made in memo_dir)}

\item{evict}{the cache eviction strategy to be used (default: "lru")}

\item{...}{additional arguments to be passed to cachem::cache_disk}
}
\value{
a memoised version of the input function,
with the cache written to disk in the specified folder.
}
\description{
The function memoise_fun_in_proj takes in a function func and returns a memoised version of the specified function,
with the cache written to disk in a specified folder within the project folder.
}
\examples{
lm_memo <- memoise_fun_in_proj(lm)
}
