% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_args_interactive.R
\name{omit_args_interactive}
\alias{omit_args_interactive}
\title{Omit Arguments Interactively}
\usage{
omit_args_interactive(func)
}
\arguments{
\item{func}{A function.}
}
\value{
A character vector with the selected argument names.
If no arguments are selected or the function has no arguments, returns \code{NULL}.
}
\description{
This function allows the user to interactively select one or more arguments
of a given function by letter.
}
\examples{

my_function <- function(arg1, arg2, arg3) {
  return(c(arg1, arg2, arg3))
}

omit_args_interactive(my_function)

# Output:
# a. arg1
# b. arg2
# c. arg3
# Choose one or more arguments to omit.
# Enter the argument letter(s) or none:
# a
# [1] "arg1"

omit_args_interactive(function(a, b) {})

# Output:
# a. a
# b. b
# Choose one or more arguments to omit.
# Enter the argument letter(s) or none:
# ab
# [1] "a" "b"

omit_args_interactive(function() {})

# Output:
# [1] NULL
}
